Refactoring main package to com.vaadin.flow.ai
