--- src/test/java/org/vaadin/addons/ai/formfiller/AddonIT.java
@@ -1,16 +1,0 @@
-package org.vaadin.addons.ai.formfiller;
-
-import com.vaadin.flow.component.html.testbench.DivElement;
-import org.junit.Assert;
-import org.junit.Test;
-
-
-public class AddonIT extends AbstractViewTest {
-
-    @Test
-    public void addonTextIsRendered() {
-        DivElement divElement = $(DivElement.class).id("theAddon");
-        Assert.assertNotNull(divElement);
-        Assert.assertEquals("Hello", divElement.getText());
-    }
-}
