--- src/test/java/org/vaadin/addons/ai/formfiller/data/ReceiptItem.java
@@ -1,33 +1,0 @@
-package org.vaadin.addons.ai.formfiller.data;
-
-import java.time.LocalDate;
-
-public class ReceiptItem {
-    private String itemQuantity;
-    private String itemDescription;
-    private Double itemCost;
-
-    public String getItemQuantity() {
-        return itemQuantity;
-    }
-
-    public void setItemQuantity(String itemQuantity) {
-        this.itemQuantity = itemQuantity;
-    }
-
-    public String getItemDescription() {
-        return itemDescription;
-    }
-
-    public void setItemDescription(String itemDescription) {
-        this.itemDescription = itemDescription;
-    }
-
-    public Double getItemCost() {
-        return itemCost;
-    }
-
-    public void setItemCost(Double itemCost) {
-        this.itemCost = itemCost;
-    }
-}
